+++ liqwid-plutarch-extra/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
+## 3.21.4 --- 2023-05-04
+
+### Added
+
+- `passetClassDataValueOf` and `passetClassDataValueOfT` to look up the amount 
+  of a value corresponding to a `PAssetClassData`.
+
+- `phasOneTokenOfAssetClassData` function for checking that a `PValue` contains
+  exactly one of a `PAssetClassData`.
+
 ## 3.21.3 -- 2023-03-28
 
 - Add `PTryFrom PData` instance for `PAssetClassData`